Split API into its own module.
