+++ package.yaml
@@ -31,8 +31,15 @@ dependencies:
 - warp
 - warp-tls
 
+library:
+  source-dirs: lib
+  exposed-modules:
+    - Slurp.Registry.API
+
 executables:
   slurp-registry-exe:
     main: Main.hs
     source-dirs: app
     ghc-options: [-threaded, -rtsopts, -with-rtsopts=-N]
+    dependencies:
+      - slurp-registry
